refactor: consolidate api endpoints
